feat(course): implement batch fetching for enrolled status and user progress
- Refactor CourseService to calculate progress percentage safely.
- Apply IN-clause batch fetching to resolve N+1 queries in getCourseList.
- Map user personalized data (enrolled, progress) efficiently using HashMap.
- Ensure clean code architecture by extracting isolated helper methods.

+++ src/main/java/com/thanhmila/codelearning/dto/response/CourseListItemResponse.java
@@ -31,6 +31,6 @@ public class CourseListItemResponse {
 
     Boolean enrolled;
 
-    Double progressPercentage;
+    Integer progressPercentage;
 
 }
